Improve generation's code organisation
