+++ requirements.txt
@@ -4,4 +4,5 @@ autopep8
 colorama
 requests
 psd-tools
+matplotlib
 python-dotenv
